requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
anthropic==0.3.11
python-dotenv==1.0.0
schedule==1.2.0